Split Utils into .h and .cpp
